WIP rip out `prefered_` functions
The `prefered_` family of functions implement the relay selector
strategy outlined in `relay-selector.md`. They take care of merging the
relay constraints which the relay selector wants to use with the user
relay constraints, taking caution to always preserve user preferences.

The semantics of merging two relay strategies in a way that always
prefer user settings is defined by the new `Intersection`, which means
that we can simply remove `prefered_` functions.